Move MetaAddr serialization into zebra_network::protocol::external
Preparation for `addrv2` serialization.
