+++ zebra-network/src/meta_addr/tests.rs
@@ -1,4 +1,3 @@
 mod check;
-mod preallocate;
 mod prop;
 mod vectors;
